Adding proper sorting to files & directories
Because we're using a stack to iterate in a depth-first way, we
need to sort paths in descending lexicographic order so that they
are processed in the correct order.

Files within the directory aren't processed from a stack, though,
so they should be sorted normally.

Note to self, could probably go back to having a single list and
process twice--once ascending to print files, and once descending
to push directories onto the paths stack. Maybe will simplify the
code?