Adding support for non-optional types
Use `flagOpt(...)` and `boolFlagOpt(...)` for when you need to detect
whether or not an argument was specified (eg. there's no suitable
default). Takes in a pointer to an optional that can be used for
testing.

Use `flag(...)` and `boolFlag(...)` for when you know a good default to
fall back on, and don't care if the user omitted it on the commandline.
Takes in a pointer to a `bool` or `[]const u8` (respectively) and will
overwrite the value when an argument is found.